feat(sequencer): embed retry logic in builder
